cloud/txt/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.8.4
gunicorn
python-dotenv==1.0.1
waitress